Issue #2354 Fix - qiskit QFT gates error during conversion (#2404)
* Fixed issue #2354 where a fallback was addded to conversions.py to decompose a qiskit circuit to native gates if an error occured. A test with zne was added to test for this with the QFT gate.

* added myself to the bottom of the authors file

* Adjusted code to follow the comments that cosenal placed with previous pull request.
Moved the exception catch from the from_qiskit() to the from_qasm() function as that function is
more general and is called by from_qiskit().
Used a QasmException catch instead of Exception to make it more clear that Qasm has trouble converting
Adjusted code to follow the comments that cosenal placed with previous pull request.
Moved the exception catch from the from_qiskit() to the from_qasm() function as that function is
more general and is called by from_qiskit().
Used a QasmException catch instead of Exception to make it more clear that Qasm has trouble converting
different gates. Removed choosing a basis for the transpile and let qiskit decide what gates to use.
Added an additional test case to test the from_qasm() function.

* resolving second round of changes reviewed by cosenal! moved the exception catch to the from_qiskit() function from the from_qasm() function as from_qasm() can handle non-qiskit circuits. Switched the test case to test the from_qiskit() function in doing so. Also only decomposed rather than transpiling and decomposing as transpiling doesn't do anything.

* renaming and added updated docstrings according to cosenal suggestions